Update code documentation per ICS372 Guidelines
I updated most of the code to conform to our instructors guidelines, but stopped at the UnserInterface class. According to his guidelines (and the person who wrote the book we are reading, who coincidentally works at this college as an instructor and was my last semester programming instructor), we are supposed to put a lot of ugly documentation code on every single method we do. We are supposed to include @param and @return on every single method, even if they have no parameters and return nothing, we are supposed to just write "nothing" to indicate that. This honestly looks disgusting to me and I think it just clogs up the code, so I didn't modify the UserInteface classes. Up to you guys if you want to keep these documentation changes or revert back, I think the older documentation looks way cleaner, but if the instructor actually wants us to do the ugly method then we don't have a choice. Should we ask him about this over email to see? I'm really not a fan of cluttering up code with useless stuff, but if he's going to dock us points if we don't then we don't really have a choice.