Move CreditCard prompt back to UserInterface from Theater
While CreditCard is a type we've introduced, which makes it seem like it should be initialised in Theater instead of UserInterface, I would treat it as a more fundamental data storage type (or, alternatively, like a C-style struct) that can be used freely in UserInterface. In contrast, a class that implements new functionality and methods, like Customer, should generally only be used inside of Theater, not inside of UserInterface.

More importantly, though, we really want to keep the Theater methods entirely free of any input and output; no prompts, no Scanners, no println()s. Why we want this becomes more apparent when you look at TestSuite, which I'll be comitting shortly.