More logical CreditCardList implementation
Basically, a Map<Long, CreditCard> instead of Map<CreditCard, Customer>. I think this works better, because in general we want to search for a full credit card object knowing only its number, in order to detect duplicates.

I'm also implementing Identifiable<Long> in CreditCard to do the dirty work of this is a more understandable way.

Finally, I added a getCreditCard(long) method to Customer that uses this new structure to perform a more optimised search of a credit card in a user's credit card list.