Serialization bug persists + Add status code to catch possible errors
The serialization bug appears to still exist. I'm looking into the other solutions you mentioned to see if those happen to fix it, otherwise it may be a problem with the credit card ArrayList getting populated? I'll mess around a bit and see.

Other than that, this is just a very minor update for error checking.