Implement minor features required on rubric
I went around and implemented a couple minor features we were missing that the rubric stated we should add in this iteration. It appears that everything functions now as it should, the only thing we have left to do is some cleanup here and there, and moving some logic out of the UserInterface and into Theater.