Very minor output change + additional changes incoming
I was reading through the rubric one last time to ensure we weren't missing anything, and it looks like we need to add only one additional thing.
"Every ticket has a serial number, the date of the show, the type of ticket (regular, advance, student advance), and the ticket price; for student advance tickets, there is an extra message that says “Must show valid student id.”"

I've implemented the "Must show valid student id." part, but it looks like when we display ticket information, we should also be outputting a serial number associated to said ticket.